the-bend-backend/pyproject.toml
[tool.poetry]
name = "the-bend-backend"
version = "0.1.0"
description = "The Bend Community App - Backend API"
authors = []
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.110.0"
uvicorn = {extras = ["standard"], version = ">=0.29.0"}
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.0"}
asyncpg = ">=0.29.0"
alembic = ">=1.13.0"
redis = {extras = ["hiredis"], version = ">=5.0.0"}
celery = {extras = ["redis"], version = ">=5.3.0"}
python-jose = {extras = ["cryptography"], version = ">=3.3.0"}
passlib = {extras = ["bcrypt"], version = ">=1.7.4"}
pywebpush = ">=2.0.0"
pydantic-settings = ">=2.2.0"
python-multipart = ">=0.0.9"
boto3 = ">=1.34.0"
httpx = ">=0.27.0"
Pillow = ">=10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.23.0"
factory-boy = ">=3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
